parsing output

+++ gather_all_output_params.py
@@ -76,7 +76,7 @@ dtype=['S8',\
            float,float,float,float,float,float,float,float,float,float,\
            'bool','bool']
 
-outtab = Table(np.zeros((nvirgo,len(dtype)),dtype=dtype),dtype=dtype,names=colnames)
+outtab = Table(np.zeros((nvirgo,len(dtype))),dtype=dtype,names=colnames)
